Move `JavadocText` to `io.spine.code.javadoc`
